+++ R/psis.R
@@ -302,9 +302,9 @@ enough_tail_samples <- function(tail_len, min_len = 5) {
 #' @return Nothing, just possibly throws warnings.
 #'
 throw_pareto_warnings <- function(k, high = 0.5, too_high = 0.7) {
-  if (any(k > too_high)) {
+  if (isTRUE(any(k > too_high))) {
     .warn("Some Pareto k diagnostic values are too high. ", .k_help())
-  } else if (any(k > high)) {
+  } else if (isTRUE(any(k > high))) {
     .warn("Some Pareto k diagnostic values are slightly high. ", .k_help())
   }
 }
@@ -363,7 +363,7 @@ prepare_psis_r_eff <- function(r_eff, len) {
   } else if (anyNA(r_eff)) {
     stop("Can't mix NA and not NA values in 'r_eff'.", call. = FALSE)
   }
-  return(r_eff)
+  r_eff
 }
 
 #' Check if `psis()` was called from one of the loo methods