+++ tests/testthat/test_loo_subsampling.R
@@ -344,6 +344,16 @@ test_that("elpd_loo_approximation works as expected", {
   fake_posterior <- draws <- as.matrix(rbeta(S, a, b))
   fake_data <- data.frame(y,K)
   rm(N, K, S, a0, b0, p, y, a, b)
+
+  list(fake_posterior = fake_posterior, fake_data = fake_data)
+}
+
+test_elpd_loo_approximation <- function(cores) {
+  set.seed(123)
+  test_data <- geterate_test_elpd_dataset()
+  fake_posterior <- test_data$fake_posterior
+  fake_data <- test_data$fake_data
+
   llfun_test <- function(data_i, draws) {
     dbinom(data_i$y, size = data_i$K, prob = draws, log = TRUE)
   }
@@ -438,9 +448,9 @@ test_that("waic using delta method and gradient", {
     # fake_posterior <- mvtnorm::rmvnorm(n = S, mean = mu_n, sigma = solve(Lambda_n))
     colnames(fake_posterior) <- c("a", "b")
     fake_data <- data.frame(y, X)
-    save(fake_posterior, fake_data, file = test_path("normal_reg_waic_test_example.rda"))
+    save(fake_posterior, fake_data, file = test_path("data-for-tests/normal_reg_waic_test_example.rda"))
   } else {
-    load(file = test_path("normal_reg_waic_test_example.rda"))
+    load(file = test_path("data-for-tests/normal_reg_waic_test_example.rda"))
   }
 
   .llfun <- function(data_i, draws) {
@@ -492,9 +502,9 @@ test_that("waic using delta 2nd order method", {
     fake_posterior[,"sigma2"] <- sqrt(fake_posterior[,"sigma2"])
     colnames(fake_posterior) <- c("a", "b", "sigma")
     fake_data <- data.frame(y, X)
-    save(fake_posterior, fake_data, file = test_path("normal_reg_waic_test_example2.rda"), compression_level = 9)
+    save(fake_posterior, fake_data, file = test_path("data-for-tests/normal_reg_waic_test_example2.rda"), compression_level = 9)
   } else {
-    load(file = test_path("normal_reg_waic_test_example2.rda"))
+    load(file = test_path("data-for-tests/normal_reg_waic_test_example2.rda"))
   }
 
   .llfun <- function(data_i, draws) {
@@ -717,7 +727,7 @@ context("loo_subsampling cases")
 test_that("Test loo_subsampling and loo_approx with radon data", {
   skip_on_cran() # avoid going over time limit for tests
 
-  load(test_path("test_radon_laplace_loo.rda"))
+  load(test_path("data-for-tests/test_radon_laplace_loo.rda"))
   # Rename to spot variable leaking errors
   llfun_test <- llfun
   log_p_test <- log_p
@@ -890,10 +900,10 @@ test_that("Test the vignette", {
          stan_df, stan_df2,
          parameter_draws, parameter_draws_laplace, parameter_draws_2,
          log_p, log_g,
-         file = test_path("loo_subsample_vignette.rda"), compression_level = 9)
+         file = test_path("data-for-tests/loo_subsample_vignette.rda"), compression_level = 9)
 
   } else {
-    load(test_path("loo_subsample_vignette.rda"))
+    load(test_path("data-for-tests/loo_subsample_vignette.rda"))
   }
 
   set.seed(4711)